+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 0.3.3
   * More consistent response with `get_typeforms()`.
   * Bug fix: Return an empty data frame when there are no typeforms (very edge case).
+  * Bug fix: typeform API insists on integers for times; so use times.
 
 # Version 0.3.2
   * Bug fix: Return an empty data frame when there are no completed responses.